Click box and move cube
When the box is clicked, the cube at the top of the box moves up and the box click event is locked. When click on another box, the cube at the top of this box moves up and the cube at the top of the previous box moves down and returns to its previous place.